feat: restructured source files
