+++ step2/console/ansi_console.c
@@ -1,18 +1,18 @@
 #include <stdio.h>
-#include "ansi.h"
+#include "board_drawer.h"
 
 /**
  * Gets terminal width and height and assign them
  * to the appropriate memory addresses
- * @param width - pointer to an integer
- * @param height - pointer to an integer 
+ * @param width address of an integer
+ * @param height address of an integer 
  * @return void
 */
 void get_terminal_size(int *width, int *height)
 {
     // Clear entire screen
     printf("\033[2J");
-    puts("Enter to start the game");
+    puts("Enter twice to start the game");
 
     // Puts cursor to the right bottom of the terminal
     printf("\033[30m\033[10000000;10000000H");
@@ -20,18 +20,13 @@ void get_terminal_size(int *width, int *height)
     printf("\033[6n");
 
     // Gets the cursor position
-    scanf("\033[%d;%dR", height, width);
+    int returned = scanf("\033[%d;%dR", height, width);
 
     // Cursor is half width of the block, thus to get the width
-    // we divide it by two
-    *width = *width / 2;
+    // we divide it by two and in case of odd number we add 1
+    *width = *width / 2 + *width % 2;
 }
 
-/**
- * Prints the game's board on the terminal
- * @param game - Pointer to the game object
- * @return void
-*/
 void draw_board(Game *game)
 {
     if (!game)
