+++ step2/console/board_drawer.h
@@ -1,0 +1,9 @@
+#ifndef __BOARD_DRAWER_H__
+#define __BOARD_DRAWER_H__
+#include "life.h"
+
+void get_terminal_size(int *width, int *height);
+
+void draw_board(Game *game);
+
+#endif
